backend/requirements.txt
fastapi==0.133.0
uvicorn==0.41.0
requests==2.32.5
yahooquery==2.4.1
cachetools==7.0.1
httpx==0.28.1
pandas
numpy
yfinance
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
lxml
beautifulsoup4
feedparser
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
anyio==4.12.1
certifi
idna
h11
typing_extensions
typing-inspection
urllib3
websockets
wrapt
multitasking
python-dateutil
pytz
six
platformdirs
packaging
protobuf
pycparser
cffi
curl_cffi
annotated-doc
annotated-types
click
Deprecated
frozendict
limits
peewee
slowapi
soupsieve
sgmllib3k
requests-futures
tqdm
